to_number: Make sure that the correct overload of pow() is chosen
[why]
The standard guarantees the overload float std::pow(float, int),
but in our previous implementation, the second parameter was
a size_t. On VS2019, this chose an overload returning double,
which in turn generated a warning because a double was
assigned to a float variable.

[how]
Add a static_cast from size_t to int.

+++ include/gul/to_number.h
@@ -163,7 +163,8 @@ constexpr inline optional<NumberType> to_unsigned_float(gul::string_view str) no
         }
         else
         {
-            digit_value = std::pow(NumberType(10), exponent + str_before_point.size() - 1);
+            digit_value = std::pow(NumberType(10),
+                    exponent + static_cast<int>(str_before_point.size()) - 1);
 
             for (char c : str_before_point)
             {
